Version 3.0.1 ======== * Fixed bug in ReflectionTool.populateFromMap(...), which needed to covert the setter property to an Object[] (and in the unit test, which was passing an Object[] instead of an object). * Fleshed out missing JavaDoc property descriptions. * Performed a bunch of micro-optimizations suggested by my IDE (remove unnecessary init assignments, inline instead of using temp variables, use StringBuilder instead of StringBuffer, use foreach, etc.)

+++ pojava-3.x/branches/pojava-3.0.1/src/main/java/org/pojava/transformation/BooleanIntegerAdaptor.java
@@ -38,7 +38,7 @@ public class BooleanIntegerAdaptor extends BindingAdaptor<Boolean, Integer> {
         if (fromBinding == null || fromBinding.getObj() == null) {
             return toBinding;
         }
-        toBinding.setObj(fromBinding.getValue().intValue() != 0);
+        toBinding.setObj(fromBinding.getValue() != 0);
         return toBinding;
     }
